requirements.txt
psutil
python-telegram-bot
pillow
